+++ src/Scripts/GetHeadersAPI.ts
@@ -1,9 +1,9 @@
 import { mhaStrings } from "./mhaStrings";
 import { Errors } from "./Errors";
-import { ParentFrame } from "./parentFrame";
+import { ParentFrame } from "./ParentFrame";
 import { GetHeaders } from "./GetHeaders";
 import { GetHeadersRest } from "./GetHeadersRest";
-import { diagnostics } from "./diag";
+import { diagnostics } from "./Diag";
 
 /*
  * GetHeadersAPI.js
