+++ src/Scripts/GetHeadersEWS.ts
@@ -1,7 +1,7 @@
 import $ from "jquery";
 import { mhaStrings } from "./mhaStrings";
 import { Errors } from "./Errors";
-import { ParentFrame } from "./parentFrame";
+import { ParentFrame } from "./ParentFrame";
 import { GetHeaders } from "./GetHeaders";
 
 /*
