+++ src/Scripts/newMobilePaneIosFrame.ts
@@ -8,7 +8,7 @@ import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc";
 import { mhaStrings } from "./mhaStrings";
 import { HeaderModel } from "./HeaderModel";
-import { Poster } from "./poster";
+import { Poster } from "./Poster";
 import { SummaryRow } from "./row/SummaryRow";
 import { Row } from "./row/Row";
 import { ReceivedRow } from "./row/ReceivedRow";
